+++ packages/localize/CHANGELOG.md
@@ -12,7 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - **[BREAKING]** Lit Localize is now distributed as two packages:
 
   - `@lit/localize` provides the browser library (`msg`, `LocalizedElement`, etc.)
-  - `@lit/localize-tools` provides the `lit-localize` binary.
+  - `@lit/localize-tools` provides the `lit-localize` CLI.
 
 ## [0.7.0] - 2021-03-12