+++ packages/localize/CHANGELOG.md
@@ -5,9 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## Unreleased
+<!-- ## Unreleased -->
 
-<!-- ### Changed -->
+## [0.8.0] - 2021-03-24
+
+### Changed
+
+- **[BREAKING]** Lit Localize is now distributed as two packages:
+
+  - `@lit/localize` provides the browser library (`msg`, `LocalizedElement`, etc.)
+  - `@lit/localize-tools` provides the `lit-localize` CLI.
 
 ## [0.7.0] - 2021-03-12